Implement `BufferBuilder` (#85)
* Implement `BufferBuilder`

* Add MATCHING_HACK

* fix lint

* fix clang-format

* run lint --fix

* change casts

* changes from reviews

* make linter happy

* changes from reviews

+++ src/4JLibraries_Source/NX/Platform.h
@@ -15,7 +15,7 @@ typedef long long LARGE_INTEGER;
 BOOL QueryPerformanceCounter(LARGE_INTEGER* lpPerformanceCount);
 
 void* XPhysicalAlloc(unsigned long, unsigned long, unsigned long, unsigned int);
-void XMemSet128(void*, int, unsigned int);
+void XMemSet128(void*, int, size_t);
 void XPhysicalFree(void*);
 
 void MemSect(int);
